fix(0158): address review findings for Phase 3 identity guards
Extract the AST-scanning helpers shared by the resource and risktypes
identity-mutation guards into internal/testutil/identitymutationguard,
removing ~330 lines of duplicated logic between the two guard test
files. Add negative assertions (no ERROR marker, unchanged risk level)
to the user_group_not_supported and no_privilege_manager dry-run
subtests so a regression that skips resolution ahead of the
privilege-support check would be caught.
